bookbroker_backend/requirements.txt
cryptography
dj-rest-auth
Django
django-allauth
django-ckeditor
django-cors-headers
django-filter
djangocms-admin-style
djangorestframework
markdown
wheel
Pillow


